Factoring out metrics and fixing autodiff fallback
Factoring out Euclidean and Riemannian metric functionality from systems
to separate objects in new `metrics` module to avoid unnecessary code
duplication and allow easier addition of new metrics.

Previously added function to fallback to autograd previously required
passing differential operator function directly which meant the
relevant functions needed to be imported to run code but this was not
possible when autograd was not available. Instead a string specifying
the name of the differential operator function is now passed, this also
allowing decoupling from the autograd backend currently used to allow
easier augmentation / replacement with an alternative backend in future.