Efficiency improvement by allowing use of cached Jacobian in some cases
Previously when evaluating gradient of log-determinant of Gram matrix in 
constrained Euclidean systems, the matrix-Hessian-product for the 
constraint function was evaluated after a corresponding call to the 
Jacobian of the constraint function at the same position value. 
Typically the Jacobian will be computed as a side-value in the MHP 
computation and cached, and so computing the MHP first and then getting 
the Jacobian should get a cache hit, preventing one unnecessary Jacobian 
evaluation.