Explicitly reassigning momentum when projecting on to cotangent space
Previously an implicit inplace update to the momentum array was applied 
when projecting on to the cotangent space. While this works for standard 
NumPy arrays, for array-like objects which don't support in place 
updates like JAX DeviceArray instances, this leads to the projection 
having no effect. The update momentum array is now explicitly reassigned 
to the state momentum attribute, meaning even if a non-inplace update 
was performed the projection will still be applied.