Always reinitialize integrator step size when adapting
Changes from previous behaviour of not reinitializing step size of
integrator if previously set on beginning of adaptive sampling stage,
which occured if for example step size was set manually when
constructing integrator or had been set in a previous adaptive
chain stage. Instead when using step size adaptation the
integrator step size is now always reinitialized to a reasonable
value when initializing the adapter.