Fixing bug with metric adaptation due to momentum not being resampled
Previously after the metric was updated at the end of an adaptive sampling stage
with a metric adapter active the chain state momentum component was left
unchanged. As under the new metric this momentum may have a very high
kinetic energy (low density) and the cached Hamiltonian and gradients depending
on this momentum are no longer valid this could cause the initialisation of
the step size in an immediately subsequent adaptive sampling stage to fail
due to the Hamiltonian at the initial state being incorrect and thus a large
chain in Hamiltonian always being computed irrespective of the step size.